skills/ai-character-consistency/lib/character_consistency.py
"""
Character Consistency Module

提供角色一致性保持功能，包括IP-Adapter、InstantID和LoRA支持。
"""

import os
import yaml
import torch
import numpy as np
from PIL import Image
from typing import Dict, List, Optional, Tuple, Union
from dataclasses import dataclass
from pathlib import Path


@dataclass
class CharacterReference:
    """角色参考数据类"""
    name: str
    reference_images: List[str]
    features: Dict
    identity_embedding: Optional[np.ndarray] = None
    style_embedding: Optional[np.ndarray] = None


class CharacterConsistency:
    """角色一致性保持主类"""
    
    def __init__(self, config_path: str = "configs/models.yaml"):
        """初始化
        
        Args:
            config_path: 配置文件路径
        """
        self.config = self._load_config(config_path)
        self.device = self._setup_device()
        self.models = {}
        
    def _load_config(self, config_path: str) -> dict:
        """加载配置"""
        if os.path.exists(config_path):
            with open(config_path, 'r') as f:
                return yaml.safe_load(f)
        return {}
    
    def _setup_device(self) -> torch.device:
        """设置计算设备"""
        if torch.cuda.is_available():
            return torch.device("cuda")
        elif hasattr(torch.backends, 'mps') and torch.backends.mps.is_available():
            return torch.device("mps")
        return torch.device("cpu")
    
    def create_reference(
        self,
        reference_images: List[str],
        name: str = "character",
        save_path: Optional[str] = None
    ) -> CharacterReference:
        """创建角色参考
        
        Args:
            reference_images: 参考图像路径列表
            name: 角色名称
            save_path: 保存路径
            
        Returns:
            CharacterReference对象
        """
        ref = CharacterReference(
            name=name,
            reference_images=reference_images,
            features=self._extract_features(reference_images),
            identity_embedding=self._extract_identity(reference_images),
            style_embedding=self._extract_style(reference_images)
        )
        
        if save_path:
            self._save_reference(ref, save_path)
            
        return ref
    
    def _extract_features(self, images: List[str]) -> Dict:
        """提取角色特征"""
        # 基础特征提取逻辑
        return {
            "hair_color": None,
            "eye_color": None,
            "age_group": None,
            "gender": None
        }
    
    def _extract_identity(self, images: List[str]) -> Optional[np.ndarray]:
        """提取身份特征"""
        # 使用Face Encoder提取
        return None
    
    def _extract_style(self, images: List[str]) -> Optional[np.ndarray]:
        """提取风格特征"""
        return None
    
    def _save_reference(self, ref: CharacterReference, path: str):
        """保存角色参考"""
        data = {
            "name": ref.name,
            "reference_images": ref.reference_images,
            "features": ref.features
        }
        with open(path, 'w') as f:
            yaml.safe_dump(data, f)
    
    def generate_pose(
        self,
        reference: CharacterReference,
        pose: str,
        num_samples: int = 1,
        output_dir: str = "output",
        controlnet_type: str = "openpose",
        strength: float = 0.8
    ) -> List[str]:
        """生成角色新姿态
        
        Args:
            reference: 角色参考
            pose: 姿态类型
            num_samples: 生成数量
            output_dir: 输出目录
            controlnet_type: ControlNet类型
            strength: 生成强度
            
        Returns:
            生成图像路径列表
        """
        os.makedirs(output_dir, exist_ok=True)
        outputs = []
        
        for i in range(num_samples):
            output_path = os.path.join(
                output_dir, 
                f"{reference.name}_{pose}_{i}.png"
            )
            # TODO: 实现实际生成逻辑
            outputs.append(output_path)
            
        return outputs
    
    def generate_expression(
        self,
        reference: CharacterReference,
        expression: str,
        num_samples: int = 1,
        output_dir: str = "output"
    ) -> List[str]:
        """生成角色新表情
        
        Args:
            reference: 角色参考
            expression: 表情类型
            num_samples: 生成数量
            output_dir: 输出目录
            
        Returns:
            生成图像路径列表
        """
        os.makedirs(output_dir, exist_ok=True)
        outputs = []
        
        for i in range(num_samples):
            output_path = os.path.join(
                output_dir,
                f"{reference.name}_{expression}_{i}.png"
            )
            outputs.append(output_path)
            
        return outputs
    
    def apply_ip_adapter(
        self,
        reference_image: str,
        target_prompt: str,
        output_path: str,
        style_scale: float = 0.8
    ) -> str:
        """应用IP-Adapter风格迁移
        
        Args:
            reference_image: 参考图像
            target_prompt: 目标提示词
            output_path: 输出路径
            style_scale: 风格强度
            
        Returns:
            输出图像路径
        """
        # TODO: 实现IP-Adapter逻辑
        return output_path
    
    def apply_instantid(
        self,
        reference_face: str,
        target_image: str,
        identity_strength: float = 0.8
    ) -> str:
        """应用InstantID身份保持
        
        Args:
            reference_face: 参考面部图像
            target_image: 目标图像
            identity_strength: 身份强度
            
        Returns:
            输出图像路径
        """
        # TODO: 实现InstantID逻辑
        return target_image
    
    def load_reference(self, path: str) -> CharacterReference:
        """加载角色参考"""
        with open(path, 'r') as f:
            data = yaml.safe_load(f)
            
        return CharacterReference(
            name=data["name"],
            reference_images=data["reference_images"],
            features=data.get("features", {})
        )


class InstantID:
    """InstantID身份保持类"""
    
    def __init__(self, config: Optional[dict] = None):
        self.config = config or {}
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.models = {}
    
    def load_identity(self, reference_path: str) -> Dict:
        """加载身份信息"""
        return {}
    
    def generate_with_identity(
        self,
        identity: Dict,
        prompt: str,
        negative_prompt: str = "",
        num_samples: int = 1
    ) -> List[str]:
        """保持身份生成"""
        return []


class SimpleLoraTrainer:
    """简化版LoRA训练器"""
    
    def __init__(self, config: Optional[dict] = None):
        self.config = config or {}
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
    
    def prepare_data(
        self,
        images_folder: str,
        output_folder: str,
        size: int = 512
    ):
        """准备训练数据
        
        Args:
            images_folder: 原始图像文件夹
            output_folder: 输出文件夹
            size: 处理后图像尺寸
        """
        os.makedirs(output_folder, exist_ok=True)
        # TODO: 实现数据准备逻辑
    
    def train(
        self,
        base_model: str,
        output_lora: str,
        learning_rate: float = 1e-4,
        max_steps: int = 500,
        batch_size: int = 4
    ):
        """训练LoRA
        
        Args:
            base_model: 基础模型路径
            output_lora: 输出LoRA路径
            learning_rate: 学习率
            max_steps: 最大步数
            batch_size: 批次大小
        """
        # TODO: 实现训练逻辑
        pass
    
    def merge_lora(
        self,
        base_model: str,
        lora_path: str,
        output_path: str,
        scale: float = 1.0
    ):
        """合并LoRA到模型"""
        pass
