skills/ai-character-consistency/requirements.txt
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
controlnet-lora>=0.1.0
ip-adapter>=0.1.0
instantid>=0.1.0
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=10.0.0
PyYAML>=6.0
numpy>=1.24.0
scipy>=1.11.0
